# src/openapi.json
{
    "openapi": "3.0.2",
    "info": {
        "title": "My Cookery Book 2 - backend",
        "version": "0.0.2",
        "license": {
            "name": "MIT"
        },
        "description": "This is rest api application backend developed on NodeJS and ExpressJS. Writen in typescript.",
        "contact": {
            "email": "tfilosk@gmail.com",
            "name": "Tomáš Filo",
            "url": "https://github.com/tfilo/my-cookery-book-2"
        }
    },
    "servers": [
        {
            "url": "http://localhost:8080/api",
            "description": "Development"
        }
    ],
...
    "paths": {
...
        "/user": {
            "post": {
                "operationId": "createUser",
                "tags": ["user"],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateUser"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "User created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/UnauthorizedError"
                    },
                    "409": {
                        "$ref": "#/components/responses/ConstraintError"
                    },
                    "422": {
                        "$ref": "#/components/responses/UnprocessableEntity"
                    },
                    "500": {
                        "$ref": "#/components/responses/InternalServerError"
                    },
                    "503": {
                        "$ref": "#/components/responses/ServiceUnavailableError"
                    }
                }
            },
            "get": {
                "operationId": "getUsers",
                "tags": ["user"],
                "responses": {
                    "200": {
                        "description": "List of users",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/SimpleUser"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/UnauthorizedError"
                    },
                    "500": {
                        "$ref": "#/components/responses/InternalServerError"
                    }
                }
            }
        },
        "/user/profile": {
            "get": {
                "operationId": "getProfile",
                "tags": ["user"],
                "responses": {
                    "200": {
                        "description": "Profile",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Profile"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/UnauthorizedError"
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    },
                    "500": {
                        "$ref": "#/components/responses/InternalServerError"
                    }
                }
            },
            "patch": {
                "operationId": "updateProfile",
                "tags": ["user"],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateProfileRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "204": {
                        "description": "Successful profile update"
                    },
                    "401": {
                        "$ref": "#/components/responses/UnauthorizedError"
                    },
                    "422": {
                        "$ref": "#/components/responses/UnprocessableEntity"
                    },
                    "500": {
                        "$ref": "#/components/responses/InternalServerError"
                    }
                }
            }
        },
...
        "/user/{userId}": {
            "get": {
                "operationId": "getUser",
                "tags": ["user"],
                "parameters": [
                    {
                        "in": "path",
                        "name": "userId",
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "minimum": 1
                        },
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/UnauthorizedError"
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    },
                    "422": {
                        "$ref": "#/components/responses/UnprocessableEntity"
                    },
                    "500": {
                        "$ref": "#/components/responses/InternalServerError"
                    }
                }
            },
            "put": {
                "operationId": "updateUser",
                "tags": ["user"],
                "parameters": [
                    {
                        "in": "path",
                        "name": "userId",
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "minimum": 1
                        },
                        "required": true
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateUser"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "User updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/UnauthorizedError"
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    },
                    "409": {
                        "$ref": "#/components/responses/ConstraintError"
                    },
                    "422": {
                        "$ref": "#/components/responses/UnprocessableEntity"
                    },
                    "500": {
                        "$ref": "#/components/responses/InternalServerError"
                    },
                    "503": {
                        "$ref": "#/components/responses/ServiceUnavailableError"
                    }
                }
            },
            "delete": {
                "operationId": "deleteUser",
                "tags": ["user"],
                "parameters": [
                    {
                        "in": "path",
                        "name": "userId",
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "minimum": 1
                        },
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "User deleted"
                    },
                    "401": {
                        "$ref": "#/components/responses/UnauthorizedError"
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    },
                    "422": {
                        "$ref": "#/components/responses/UnprocessableEntity"
                    },
                    "500": {
                        "$ref": "#/components/responses/InternalServerError"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        },
        "schemas": {
...
            "UpdateProfileRequest": {
                "type": "object",
                "required": ["password", "newPassword", "firstName", "lastName", "email", "notifications"],
                "properties": {
                    "password": {
                        "type": "string",
                        "format": "password",
                        "maxLength": 255,
                        "description": "Old password to verify this user"
                    },
                    "newPassword": {
                        "type": "string",
                        "format": "password",
                        "nullable": true,
                        "minLength": 8,
                        "maxLength": 255,
                        "description": "New password for user. Should be at least 8 characters long, at least one small letter, one capital letter and one number."
                    },
                    "firstName": {
                        "type": "string",
                        "minLength": 3,
                        "maxLength": 50,
                        "nullable": true
                    },
                    "lastName": {
                        "type": "string",
                        "minLength": 3,
                        "maxLength": 50,
                        "nullable": true
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "maxLength": 320
                    },
                    "notifications": {
                        "type": "boolean"
                    }
                }
            },
...
            "SimpleUser": {
                "type": "object",
                "required": ["id", "username", "firstName", "lastName", "confirmed", "notifications", "roles"],
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int32",
                        "minimum": 1
                    },
                    "username": {
                        "type": "string",
                        "minLength": 1,
                        "maxLength": 50
                    },
                    "firstName": {
                        "type": "string",
                        "minLength": 1,
                        "maxLength": 50,
                        "nullable": true
                    },
                    "lastName": {
                        "type": "string",
                        "minLength": 1,
                        "maxLength": 50,
                        "nullable": true
                    },
                    "confirmed": {
                        "type": "boolean"
                    },
                    "notifications": {
                        "type": "boolean"
                    },
                    "roles": {
                        "type": "array",
                        "items": {
                            "type": "string",
                            "enum": ["ADMIN", "CREATOR"]
                        }
                    }
                }
            },
            "CreateUser": {
                "type": "object",
                "required": ["username", "password", "firstName", "lastName", "email", "notifications", "roles"],
                "properties": {
                    "username": {
                        "type": "string",
                        "minLength": 4,
                        "maxLength": 50,
                        "description": "Only a-z 0-9 characters allowed"
                    },
                    "password": {
                        "type": "string",
                        "format": "password",
                        "minLength": 8,
                        "maxLength": 255,
                        "description": "New password for user. Should be at least 8 characters long, at least one small letter, one capital letter and one number."
                    },
                    "firstName": {
                        "type": "string",
                        "minLength": 3,
                        "maxLength": 50,
                        "nullable": true
                    },
                    "lastName": {
                        "type": "string",
                        "minLength": 3,
                        "maxLength": 50,
                        "nullable": true
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "maxLength": 320
                    },
                    "notifications": {
                        "type": "boolean"
                    },
                    "roles": {
                        "type": "array",
                        "items": {
                            "type": "string",
                            "enum": ["ADMIN", "CREATOR"]
                        }
                    }
                }
            },
...
            "Profile": {
                "type": "object",
                "required": ["username", "firstName", "lastName", "email", "confirmed", "notifications"],
                "properties": {
                    "username": {
                        "type": "string",
                        "minLength": 1,
                        "maxLength": 50
                    },
                    "firstName": {
                        "type": "string",
                        "minLength": 1,
                        "maxLength": 50,
                        "nullable": true
                    },
                    "lastName": {
                        "type": "string",
                        "minLength": 1,
                        "maxLength": 50,
                        "nullable": true
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "maxLength": 320
                    },
                    "confirmed": {
                        "type": "boolean"
                    },
                    "notifications": {
                        "type": "boolean"
                    }
                }
            },
            "User": {
                "type": "object",
                "required": [
                    "id",
                    "username",
                    "firstName",
                    "lastName",
                    "email",
                    "confirmed",
                    "notifications",
                    "roles",
                    "createdAt",
                    "updatedAt"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int32",
                        "minimum": 1
                    },
                    "username": {
                        "type": "string",
                        "minLength": 1,
                        "maxLength": 50
                    },
                    "firstName": {
                        "type": "string",
                        "minLength": 1,
                        "maxLength": 50,
                        "nullable": true
                    },
                    "lastName": {
                        "type": "string",
                        "minLength": 1,
                        "maxLength": 50,
                        "nullable": true
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "maxLength": 320
                    },
                    "confirmed": {
                        "type": "boolean"
                    },
                    "notifications": {
                        "type": "boolean"
                    },
                    "roles": {
                        "type": "array",
                        "items": {
                            "type": "string",
                            "enum": ["ADMIN", "CREATOR"]
                        }
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2022-04-10T12:58:33.592Z"
                    },
                    "updatedAt": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2022-04-10T12:58:33.592Z"
                    }
                }
            },
...
        },
        "responses": {
            "UnauthorizedError": {
                "description": "Access token is missing or invalid. Possibly incorrect credentials.",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/Error"
                        }
                    }
                }
            },
            "NotFound": {
                "description": "Object is missing, or required parent doesn't exists.",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/Error"
                        }
                    }
                }
            },
            "InternalServerError": {
                "description": "General error. Something went wrong on server side.",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/Error"
                        }
                    }
                }
            },
            "UnprocessableEntity": {
                "description": "Request can't be processed because validation failed.",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/Error"
                        }
                    }
                }
            },
            "ServiceUnavailableError": {
                "description": "Request can't be processed because mail service failed.",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/Error"
                        }
                    }
                }
            },
            "ConstraintError": {
                "description": "Constraint or duplicity error.",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/Error"
                        }
                    }
                }
            },
...
        }
    },
    "security": [
        {
            "bearerAuth": []
        }
    ]
}